Implement Phase 2 concurrency and refactor database session logic
- Implemented blocking locks and `SET LOCK_TIMEOUT` (Phase 2).
- Refactored monolithic `session.rs` into `analyzer.rs`, `execution.rs`, and `dispatch.rs` following SRP.
- Fixed lock leakage in autocommit mode for read-only statements.
- Added comprehensive integration tests in `concurrency_phase2.rs`.
- Ensured all existing tests pass and fixed unused import warnings.
