+++ crates/tsql_core/src/executor/database/mod.rs
@@ -1,4 +1,6 @@
-pub(crate) mod session;
+pub(crate) mod analyzer;
+pub(crate) mod execution;
+pub(crate) mod dispatch;
 pub(crate) mod engine;
 pub(crate) mod persistence;
